Build autosuggest input that filters test results
Add an input box in the header that autosuggests available directories
to filter test results against. Helpful to concentrate on seeing what
has failed in a particular vertical or portal.

You can also filter on any string.

It is based on this obscure completely.js autosuggest library
http://complete-ly.appspot.com/.

Had to modify completely.js source a tiny bit. This library doesn’t
work with bower.