Add ability to bless test images clientside using POST
New /baseline post route added to server to kick off a process that
moves the ‘blessed’ test file into the bitmaps_reference folder thus
making it a new baseline. It also updates the compare config.json that
serves as the data for the client compare page. This allows us to
update the testStatus as ‘blessed’ on the client without reloading the
page.