Add faux checkboxes to fix bug
For some reason, an input control in the sticky header was causing them
to flip out if you clicked on them. Very weird. They would cause the
page to scroll to the top whenever a sticky header event fired.

Was going to update the style of the checkboxes anyway, not exactly
what I had in mind, but it works.